requirements.txt
aio_pika
psycopg2-binary
python-dotenv
dataclasses; python_version < '3.7'
langchain-core
langchain-openai